% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jev.R
\name{jev_api_key}
\alias{jev_api_key}
\title{Read the API key from the environment}
\usage{
jev_api_key()
}
\value{
The value of the \code{TYPESAFE_API_KEY} environment variable, or an
empty string if it is not set.
}
\description{
Read the API key from the environment
}
