% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers.R
\name{jev_usage}
\alias{jev_usage}
\title{Token usage of a response}
\usage{
jev_usage(res)
}
\arguments{
\item{res}{A \code{jev_response} from \code{\link[=jev]{jev()}}.}
}
\value{
A one-row tibble with \code{model}, \code{input_tokens} and \code{output_tokens}.
}
\description{
Token usage of a response
}
\seealso{
Other answers:
\code{\link[=jev_answer]{jev_answer()}},
\code{\link[=jev_probabilities]{jev_probabilities()}},
\code{\link[=jev_tidy]{jev_tidy()}}
}
\concept{answers}
